pyproject.toml
[tool.poetry]
name = "model-deployment-app"
version = "2.3.0"
description = ""
authors = ["YankaiW <wyk.uestc@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
fastapi = "^0.74.1"
uvicorn = "^0.17.5"
PyYAML = "^6.0"
google-cloud-storage = "^2.6.0"
scikit-learn = "1.1.1"
torch = { url = "https://download.pytorch.org/whl/cpu/torch-1.12.1%2Bcpu-cp39-cp39-linux_x86_64.whl" }

[tool.poetry.dev-dependencies]
pytest = "^7.0.1"
isort = "^5.10.1"
flake8 = "^4.0.1"
black = "^22.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
